Refactor class names
