+++ app/src/main/java/com/gnoemes/bubblenotes/data/source/DataManager.java
@@ -1,32 +1,13 @@
 package com.gnoemes.bubblenotes.data.source;
 
-
 import com.gnoemes.bubblenotes.data.model.Note;
-import com.gnoemes.bubblenotes.data.source.remote.IRemoteDataManager;
-
-import javax.inject.Inject;
-import javax.inject.Singleton;
 
 import io.realm.RealmResults;
 
 /**
- * DataManager which controls Realm Database, API
+ * Created by kenji1947 on 25.09.2017.
  */
 
-@Singleton
-public class DataManager implements Manager.DataManager {
-
-    private final Manager.RealmManager realmManager;
-    private final IRemoteDataManager remoteDataManager;
-
-    @Inject
-    public DataManager(Manager.RealmManager realmManager, IRemoteDataManager remoteDataManager) {
-        this.realmManager = realmManager;
-        this.remoteDataManager = remoteDataManager;
-    }
-
-    @Override
-    public RealmResults<Note> loadNotes(Class clazz) {
-        return realmManager.loadNotes(clazz);
-    }
+public interface DataManager {
+    RealmResults<Note> loadNotes(Class clazz);
 }
