--- app/src/main/java/com/gnoemes/bubblenotes/data/source/Manager.java
@@ -1,25 +1,0 @@
-package com.gnoemes.bubblenotes.data.source;
-
-import com.gnoemes.bubblenotes.data.model.Note;
-
-import io.realm.RealmResults;
-
-public interface Manager {
-
-    /**
-     * Realm Database Manager interface
-     */
-    interface RealmManager {
-
-        RealmResults<Note> loadNotes(Class clazz);
-    }
-
-    /**
-     * Manages all model-related issues
-     */
-
-    interface DataManager {
-
-        RealmResults<Note> loadNotes(Class clazz);
-    }
-}
