+++ app/src/main/java/com/gnoemes/bubblenotes/data/source/local/RealmManager.java
@@ -1,25 +1,13 @@
 package com.gnoemes.bubblenotes.data.source.local;
 
-import com.gnoemes.bubblenotes.App;
-import com.gnoemes.bubblenotes.data.source.Manager;
 import com.gnoemes.bubblenotes.data.model.Note;
-
-import javax.inject.Inject;
 
 import io.realm.RealmResults;
 
-public class RealmManager implements Manager.RealmManager {
+/**
+ * Created by kenji1947 on 25.09.2017.
+ */
 
-    @Inject
-    RealmDatabase realmDatabase;
-
-    public RealmManager() {
-        App.getAppComponent().inject(this);
-    }
-
-
-    @Override
-    public RealmResults<Note> loadNotes(Class clazz) {
-        return realmDatabase.find(clazz);
-    }
+public interface RealmManager {
+    RealmResults<Note> loadNotes(Class clazz);
 }
