--- app/src/main/java/com/gnoemes/bubblenotes/data/source/remote/RemoteDataSource.java
@@ -1,4 +1,0 @@
-package com.gnoemes.bubblenotes.data.source.remote;
-
-public class RemoteDataSource implements IRemoteDataManager {
-}
