+++ app/src/main/java/com/gnoemes/bubblenotes/data/source/remote/RemoteManager.java
@@ -1,4 +1,4 @@
 package com.gnoemes.bubblenotes.data.source.remote;
 
-public interface IRemoteDataManager {
+public interface RemoteManager {
 }
