+++ app/src/main/java/com/gnoemes/bubblenotes/data/source/remote/RemoteManagerDefault.java
@@ -1,0 +1,4 @@
+package com.gnoemes.bubblenotes.data.source.remote;
+
+public class RemoteManagerDefault implements RemoteManager {
+}
